[FEAT] 예외처리와 에러코드 처리 로직 추가

+++ src/main/java/assignment/wanted_pre_onboarding/service/RecruitmentService.java
@@ -1,17 +1,20 @@
 package assignment.wanted_pre_onboarding.service;
 
+import assignment.wanted_pre_onboarding.common.exception.model.NotFoundException;
 import assignment.wanted_pre_onboarding.domain.company.Company;
 import assignment.wanted_pre_onboarding.domain.company.CompanyRepository;
 import assignment.wanted_pre_onboarding.domain.recruitment.Recruitment;
 import assignment.wanted_pre_onboarding.domain.recruitment.RecruitmentRepository;
-import assignment.wanted_pre_onboarding.web.recruitment.dto.request.*;
-import assignment.wanted_pre_onboarding.web.recruitment.dto.response.*;
+import assignment.wanted_pre_onboarding.controller.recruitment.dto.request.*;
+import assignment.wanted_pre_onboarding.controller.recruitment.dto.response.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+
+import static assignment.wanted_pre_onboarding.common.exception.ErrorCode.NOT_FOUND_EXCEPTION;
 
 @Service
 @RequiredArgsConstructor
@@ -22,6 +25,10 @@ public class RecruitmentService {
 
     public CreateRecruitResponse createRecruit(CreateRecruitRequest request) {
         Optional<Company> findCompany = companyRepository.findById(request.getCompanyId());
+
+        if (findCompany.isEmpty()) {
+            throw new NotFoundException("회사 등록 후 채용공고를 등록할 수 있습니다.", NOT_FOUND_EXCEPTION);
+        }
 
         Recruitment recruitment = Recruitment.of(findCompany.get(),
                 request.getPosition(),
@@ -37,8 +44,11 @@ public class RecruitmentService {
     }
 
     public UpdateRecruitResponse updateRecruit(UpdateRecruitRequest request) {
+        Optional<Recruitment> findRecruit = recruitmentRepository.findById(request.getRecruitId());
 
-        Optional<Recruitment> findRecruit = recruitmentRepository.findById(request.getId());
+        if (findRecruit.isEmpty()) {
+            throw new NotFoundException("등록되지 않은 채용공고입니다.", NOT_FOUND_EXCEPTION);
+        }
 
         findRecruit.get().setPosition(request.getPosition());
         findRecruit.get().setReward(request.getReward());
@@ -52,13 +62,16 @@ public class RecruitmentService {
         return response;
     }
 
-    public String deleteRecruit(DeleteRecruitRequest request) {
-        Optional<Recruitment> findRecruit = recruitmentRepository.findById(request.getId());
+    public void deleteRecruit(DeleteRecruitRequest request) {
+        Optional<Recruitment> findRecruit = recruitmentRepository.findById(request.getRecruitId());
+
+        if (findRecruit.isEmpty()) {
+            throw new NotFoundException("존재하지 않는 채용공고입니다.", NOT_FOUND_EXCEPTION);
+        }
+
         String position = findRecruit.get().getPosition();
 
         recruitmentRepository.delete(findRecruit.get());
-
-        return "해당 " + position + " 채용 공고가 삭제되었습니다.";
     }
 
     public List<GetRecruitListResponse> getRecruitList() {
@@ -99,6 +112,10 @@ public class RecruitmentService {
 
     public GetRecruitDetailResponse getRecruitDetail(Long id) {
         Optional<Recruitment> findRecruit = recruitmentRepository.findById(id);
+        if (findRecruit.isEmpty()) {
+            throw new NotFoundException("존재하지 않는 채용공고입니다.", NOT_FOUND_EXCEPTION);
+        }
+
         Company company = findRecruit.get().getCompany();
         Long companyId = company.getId();
